--- components/app_jsx.txt
@@ -1,16 +1,0 @@
-// import {CreateInput} from "./createInput.jsx"
-// import {Web_pilot} from "../web_pilot/web_pilot.jsx"
-
-function App(){
-
-    return(
-
-        <CreateInput/>
-    )
-}
-
-export default App;
-
-let appendHere = document.getElementsByClassName("header")[0];
-const showItem = Web_pilot.createElement(App);
-Web_pilot.render(showItem, appendHere);
