+++ package.json
@@ -4,7 +4,7 @@
   "description": "(01F JS project #5)",
   "source": "static/index.html",
   "scripts": {
-    "start": "parcel",
+    "start": "parcel static/index.html",
     "build": "parcel build --dist-dir public"
   },
   "repository": {
